Poker games
The final product is a windows executable file that is easy to setup and play. The executable includes a game selection menu with a Three Card Poker game currently implemented. Gameplay is designed primarily for mouse interaction, and works well with touch screens. The executable uses a configured blackjack logic with the Simple DirectMedia Layer Library for graphics and input. Additional logic is used for the game menu.